+++ src/Applications/use_case/DeleteReplyCommentUseCase.js
@@ -1,17 +1,16 @@
 
 class DeleteReplyCommentUseCase {
-  constructor({replyCommentRepository}) {
+  constructor({replyCommentRepository, commentRepository}) {
     this._replyCommentRepository = replyCommentRepository;
+    this._commentRepository = commentRepository;
   }
 
   async execute(useCasePayload) {
     const {threadId, commentId, owner, replyId} = useCasePayload;
-    console.log(owner, 'DeleteReplyCommentUseCase');
-    await this._replyCommentRepository.verifyThreadId(threadId);
-    await this._replyCommentRepository.verifyCommentId(commentId);
-    await this._replyCommentRepository.verifyReply(replyId);
-    await this._replyCommentRepository.verifyOwner(owner, replyId);
-    return this._replyCommentRepository.deleteReplyComment(replyId);
+    // gimana kalo thread dan comment tidak ada tapi reply ada
+    await this._replyCommentRepository.verifyExistsCommentAndThreadByIds({threadId, commentId});
+    await this._replyCommentRepository.verifyExistsReplyById(replyId);
+    return this._replyCommentRepository.deleteReplyComment({replyId, owner});
   }
 }
